requirements.txt
python==3.9.17
numpy==1.25.1
scipy==1.11.1
scikit-learn==1.3.0
torch==2.0.1
torchvision==0.15.2
pytorch-cuda==11.7
tqdm==4.65.0
tensorboard==2.13.0
pyyaml==6.0.1
pyyaml-include==1.3.1
yacs==0.1.8
albumentations==1.4.3
